refactor: extract test-bot from echo-bot samples
Move card/invoke handlers into dedicated test-bot samples in all 3
languages. Echo bots are now minimal (message echo only). Test-bot
includes: card command, adaptiveCard/action invoke, test/echo invoke.

- Add dotnet/samples/TestBot, node/samples/test-bot, python/samples/test-bot
- Revert echo bots to minimal implementations
- Add TestBot to Botas.slnx
- Update Playwright README with test-bot instructions
- Fix Playwright invoke test: sendRawMessage + iframe fallback
- Update E2E invoke test docs
