+++ node/samples/test-bot/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "test-bot",
+  "version": "0.1.0",
+  "type": "module",
+  "scripts": {
+    "start": "node --import tsx index.ts",
+    "dev": "node --import tsx --watch index.ts"
+  },
+  "dependencies": {
+    "botas-express": "*"
+  }
+}
