+++ python/samples/test-bot/pyproject.toml
@@ -1,0 +1,7 @@
+[project]
+name = "botas-sample-test-bot"
+version = "0.1.0"
+requires-python = ">=3.11"
+dependencies = [
+    "botas-fastapi",
+]
